added logic behind adding new meal. created 1 test for it. tests W.I.P.

+++ barvBackEnd/src/main/java/com/barv/food/Food.java
@@ -44,5 +44,5 @@ public class Food {
     private double fats;
     private double calories;
     //Weight in grams
-    private double weight;
+    private int weight;
 }
